requirements.txt
anyio==3.7.0
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.99.0
h11==0.14.0
idna==3.4
pydantic==1.10.10
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.0
uvicorn==0.22.0
anyio==3.7.0
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.99.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
pydantic==1.10.10
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.0
uvicorn==0.22.0
anyio==3.7.0
asyncpg==0.27.0
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.99.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
psycopg2-binary==2.9.6
pydantic==1.10.10
sniffio==1.3.0
SQLAlchemy==2.0.17
starlette==0.27.0
typing_extensions==4.7.0
uvicorn==0.22.0
anyio==3.7.0
asyncpg==0.27.0
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.99.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
psycopg2-binary==2.9.6
pydantic==1.10.10
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.17
starlette==0.27.0
typing_extensions==4.7.0
uvicorn==0.22.0
anyio==3.7.0
asyncpg==0.27.0
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.99.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
psycopg2-binary==2.9.6
pydantic==1.10.10
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a34
sqlmodel==0.0.8
starlette==0.27.0
typing_extensions==4.7.0
uvicorn==0.22.0
